# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tripio",
        "description": "Документация API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Вывод списка записей (index())",
                "description": "Доступ только авторизованным пользователям",
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    },
                    "200": {
                        "description": "Everything OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Добавление новой записи (store())",
                "description": "Возвращает новый созданный room",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/rooms/{id}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Вывод конкретной записи (show())",
                "description": "Returns room data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Обновление записи (update())",
                "description": "Returns updated room data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoomSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Удаление записи (delete())",
                "description": "Deletes a record and returns no content",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Room id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RoomSchema": {
                "properties": {
                    "id": {
                        "description": "Room ID",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "hotel_id": {
                        "description": "Foreign key hotel",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "type": {
                        "description": "Тип комнаты",
                        "type": "enum",
                        "example": "standard double",
                        "nullable": false
                    },
                    "price": {
                        "description": "Стоимость номера",
                        "type": "number",
                        "format": "float",
                        "example": "10500.99",
                        "nullable": false
                    },
                    "status": {
                        "description": "Занят номер (0) или нет (1)",
                        "type": "tinyint",
                        "example": "0",
                        "nullable": false
                    },
                    "created_at": {
                        "description": "Date of route creation",
                        "type": "string",
                        "format": "date-time",
                        "nullable": false
                    },
                    "updated_at": {
                        "description": "Date of last updating route data",
                        "type": "string",
                        "format": "date-time",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}